+++ lib/Subscriber/Behavior/Mapping/NodeNameSubscriber.php
@@ -28,17 +28,42 @@ class NodeNameSubscriber implements EventSubscriberInterface
     public static function getSubscribedEvents()
     {
         return [
-            Events::HYDRATE => 'handleNodeName',
-            Events::PERSIST => 'handleNodeName',
+            Events::HYDRATE => 'setFinalNodeName',
+            Events::PERSIST => [
+                ['setInitialNodeName', 0],
+                ['setFinalNodeName', -480],
+            ],
         ];
     }
 
     /**
+     * Sets the initial node name.
+     *
+     * @param AbstractMappingEvent $event
+     */
+    public function setInitialNodeName(AbstractMappingEvent $event)
+    {
+        $this->setNodeName($event);
+    }
+
+    /**
+     * Sets the final node name at the end, in case it was changed.
+     *
+     * @param AbstractMappingEvent $event
+     */
+    public function setFinalNodeName(AbstractMappingEvent $event)
+    {
+        $this->setNodeName($event);
+    }
+
+    /**
+     * Sets the node name.
+     *
      * @param AbstractMappingEvent $event
      *
      * @throws DocumentManagerException
      */
-    public function handleNodeName(AbstractMappingEvent $event)
+    private function setNodeName(AbstractMappingEvent $event)
     {
         $document = $event->getDocument();