+++ tests/Unit/Subscriber/Behavior/Path/AliasFilingSubscriberTest.php
@@ -12,12 +12,12 @@
 namespace Sulu\Component\DocumentManager\Tests\Unit\Subscriber\Behavior\Audit\Path;
 
 use PHPCR\NodeInterface;
+use PHPCR\SessionInterface;
 use Sulu\Component\DocumentManager\Behavior\Path\AliasFilingBehavior;
 use Sulu\Component\DocumentManager\DocumentManager;
 use Sulu\Component\DocumentManager\Event\PersistEvent;
 use Sulu\Component\DocumentManager\Metadata;
 use Sulu\Component\DocumentManager\MetadataFactoryInterface;
-use Sulu\Component\DocumentManager\NodeManager;
 use Sulu\Component\DocumentManager\Subscriber\Behavior\Path\AliasFilingSubscriber;
 
 class AliasFilingSubscriberTest extends \PHPUnit_Framework_TestCase
@@ -28,11 +28,6 @@ class AliasFilingSubscriberTest extends \PHPUnit_Framework_TestCase
     private $persistEvent;
 
     /**
-     * @var \stdClass
-     */
-    private $notImplementing;
-
-    /**
      * @var AliasFilingBehavior
      */
     private $document;
@@ -41,11 +36,6 @@ class AliasFilingSubscriberTest extends \PHPUnit_Framework_TestCase
      * @var \stdClass
      */
     private $parentDocument;
-
-    /**
-     * @var NodeManager
-     */
-    private $nodeManager;
 
     /**
      * @var DocumentManager
@@ -68,6 +58,26 @@ class AliasFilingSubscriberTest extends \PHPUnit_Framework_TestCase
     private $parentNode;
 
     /**
+     * @var SessionInterface
+     */
+    private $defaultSession;
+
+    /**
+     * @var NodeInterface
+     */
+    private $defaultNode;
+
+    /**
+     * @var SessionInterface
+     */
+    private $liveSession;
+
+    /**
+     * @var NodeInterface
+     */
+    private $liveNode;
+
+    /**
      * @var AliasFilingSubscriber
      */
     private $subscriber;
@@ -75,17 +85,25 @@ class AliasFilingSubscriberTest extends \PHPUnit_Framework_TestCase
     public function setUp()
     {
         $this->persistEvent = $this->prophesize(PersistEvent::class);
-        $this->notImplementing = new \stdClass();
         $this->document = $this->prophesize(AliasFilingBehavior::class);
         $this->parentDocument = new \stdClass();
-        $this->nodeManager = $this->prophesize(NodeManager::class);
         $this->documentManager = $this->prophesize(DocumentManager::class);
         $this->metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
         $this->metadata = $this->prophesize(Metadata::class);
         $this->parentNode = $this->prophesize(NodeInterface::class);
+        $this->defaultSession = $this->prophesize(SessionInterface::class);
+        $this->defaultNode = $this->prophesize(NodeInterface::class);
+        $this->liveSession = $this->prophesize(SessionInterface::class);
+        $this->liveNode = $this->prophesize(NodeInterface::class);
+
+        $this->metadataFactory->getMetadataForClass(get_class($this->document->reveal()))
+            ->willReturn($this->metadata->reveal());
+        $this->defaultSession->getRootNode()->willReturn($this->defaultNode->reveal());
+        $this->liveSession->getRootNode()->willReturn($this->liveNode->reveal());
 
         $this->subscriber = new AliasFilingSubscriber(
-            $this->nodeManager->reveal(),
+            $this->defaultSession->reveal(),
+            $this->liveSession->reveal(),
             $this->metadataFactory->reveal()
         );
     }
@@ -95,7 +113,7 @@ class AliasFilingSubscriberTest extends \PHPUnit_Framework_TestCase
      */
     public function testPersistNotImplementing()
     {
-        $this->persistEvent->getDocument()->willReturn($this->notImplementing);
+        $this->persistEvent->getDocument()->willReturn(new \stdClass());
         $this->subscriber->handlePersist($this->persistEvent->reveal());
     }
 
@@ -105,15 +123,28 @@ class AliasFilingSubscriberTest extends \PHPUnit_Framework_TestCase
     public function testSetParentDocument()
     {
         $this->persistEvent->getDocument()->willReturn($this->document->reveal());
-        $this->persistEvent->getLocale()->willReturn('fr');
-        $this->metadataFactory->getMetadataForClass(get_class($this->document->reveal()))->willReturn(
-            $this->metadata->reveal()
-        );
-        $this->metadata->getAlias()->willReturn('test');
-        $this->nodeManager->createPath('/tests')->willReturn($this->parentNode->reveal());
-        $this->persistEvent->hasParentNode()->shouldBeCalled();
-        $this->persistEvent->setParentNode($this->parentNode->reveal())->shouldBeCalled();
-        $this->documentManager->find('/tests', 'fr')->willReturn($this->parentDocument);
+        $this->persistEvent->hasParentNode()->willReturn(true);
+        $this->persistEvent->getParentNode()->willReturn($this->parentNode->reveal());
+        $this->parentNode->getPath()->willReturn('/cmf');
+        $this->metadata->getAlias()->willReturn('banner');
+
+        $defaultCmfNode = $this->prophesize(NodeInterface::class);
+        $this->defaultNode->hasNode('cmf')->willReturn(true);
+        $this->defaultNode->getNode('cmf')->willReturn($defaultCmfNode->reveal());
+
+        $defaultBannerNode = $this->prophesize(NodeInterface::class);
+        $defaultCmfNode->hasNode('banners')->willReturn(false);
+        $defaultCmfNode->addNode('banners')->willReturn($defaultBannerNode->reveal());
+
+        $liveCmfNode = $this->prophesize(NodeInterface::class);
+        $this->liveNode->hasNode('cmf')->willReturn(true);
+        $this->liveNode->getNode('cmf')->willReturn($liveCmfNode->reveal());
+
+        $liveBannerNode = $this->prophesize(NodeInterface::class);
+        $liveCmfNode->hasNode('banners')->willReturn(false);
+        $liveCmfNode->addNode('banners')->willReturn($liveBannerNode->reveal());
+
+        $this->persistEvent->setParentNode($defaultBannerNode->reveal())->shouldBeCalled();
 
         $this->subscriber->handlePersist($this->persistEvent->reveal());
     }